requirements.txt
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
